dt-bindings: remoteproc: add compatibility for TEE support
Rework compatibility description according to the support of
the authenticated firmware relying on OP-TEE authentication.

the expected behavior is:
- with compatible "st,stm32mp1-m4" the Linux kernel load a non signed (ELF)
firmware
with compatible "st,stm32mp1-m4-tee" the Linux kernel load a signed
firmware. In this case it call TEE services to manage the firmware loading
and the remoteproc life-cycle.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>