dt-bindings: remoteproc: add bindings for remoteproc virtio device
The remoteproc framework can rely on virtio devices for the
inter-processor communication. In such case shared memory regions
has to be defined for the allocation of the vrings and the shared
buffers.
These bindings describe the virtio device by specifying the type of
virtio device and the associated memory regions.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>