remoteproc: virtio: create bind/unbind ops
The virtio platform driver can be instantiate via the devicetree or
by the remoteproc core itself.
A bind/unbind phase is needed to create remoteproc subdevs only if a
virtio resource is detected in the resource table.
Indeed a virtio can be declared in device tree but not used, depending
on the remote firmware. In this case the remoteproc must not start it.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>