drivers: mailbox: add interrupt controller to manage interrupt context
The mailbox client callback should be called under interrupt context or in
normal context. The need depends on the client.
For instance a virtio I2C client needs to be called under interrupt
context. Indeed the associated bus controller should allows an I2C drivers
to access to the hardware during the probe ( atomic context).
In such case the remote device "Ack" reception have to be to be treated
under interrupt context.
On the other hand a virtio rpmsg should not work in interrupt context. It
mixes several services in one virtio.

This patch add the capability to call the mailbox client under interrupt
or normal context depending on the mailbox channel ID.

- if the interrupt associated to a mailbox channel is declared in the
  client device tree node, the client is called in interrupt context
- else the client is called in normal context (use of a workqueue)

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>