virtio_mmio: Added compatiblility with remote processor
-Add support of dma region to allocate buffers,
this is necessary to ensure that the dma memory is
located in the shared memory with the coprocessor

-Add virtio,mmio-remote proc compatible wich uses,
mailboxes to send interrupts to the remote processor.
We need this because the coprocessor can't react to
register changes otherwise. And VirtIO MMIO assume that
writing to a register update instantly the value of
other registers.

Example:
VIRTIO_MMIO_DRIVER_FEATURES_SEL is used to choose
wich bits will be present in VIRTIO_MMIO_DRIVER_FEATURES.

But without interrupt, the register can't be updated
before the driver write again in the same register.

Signed-off-by: Nicolas Granger <nicolas.granger01@st.com>