rpmsg: virtio : extend sending of destroy announcement
rpdev->announce is used to prevent sending a name service (NS) announcement
when a local RPMsg device is created by the remote side using the NS
announcement service.

A side effect is that it becomes impossible to send an NS destroy
announcement when this local RPMsg device is removed.
Indeed, virtio_rpmsg_announce_destroy checks rpdev->announce, which
is false in this case.

As a result, the remote side is never informed that the service is
no longer supported.

This patch proposes to always send an NS destroy to the remote side if
VIRTIO_RPMSG_F_NS is supported and a name service is defined.

Exceptions are the rpmsg_ns and rpmsg_ctrl devices:
- rpmsg_ctrl has no endpoint
- rpmsg_ns uses the reserved RPMSG_NS_ADDR address

Change-Id: I3cd26b98416c7617eda94f32d364c093a4e26ad1
Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>